:art: move logos into logo directory
